Pdlp fix batch cuda graph (#68)
This PR aims at fixing the invalid operation while there is a graph
capture we sometimes see when using batch solve.

Solution is to use a regular instead of a non-blocking stream to make
sure that if any operation (like a cudaFree from Thrust) is being
launched on the default stream, it will wait for all other operations on
other stream to finish first, preventing any cudaMalloc/Free while
another stream might be doing a CUDA Graph capture.