Write out mps file containing user problem. Useful for debugging (#66)
This PR adds the string parameter CUOPT_USER_PROBLEM_FILE =
"user_problem_file". By default the parameter is the empty string "". If
the parameter is set by the user to something other than the empty
string, we will write out an MPS file containing the LP or MIP.

This is very useful when trying to debug interfaces like CVXPY where the
interface between CVXPY and cuOpt may have performed transformation to
the user problem. This allows us to reproduce failures on the engine
side directly from the MPS file without needing Python scripts.