Feat: rework cmakelists.txt to facilitate deb file creation. (#190)
Adds changes to the top level cmake_lists.txt to enable cpack creation of .deb files. In order to create the runtime and develop .deb files you will need to:
1) Build cuOpt
2) Navigate to cpp/build
3) run  cpack -G DEB

This creates two .deb files for the architecture of the machine that cuOpt is built on. 
For example: `cuOpt_25.08.0_amd64-Development.deb` and `cuOpt_25.08.0_amd64-Runtime.deb`

The runtime deb package contains the cuopt CLI, the mps reader.so and the cuopt.so.
The cuopt_CLI is installed in the `/usr/local/bin` directory
The .so files are installed in the `/usr/local/lib` directory

The development deb package contains the mps reader.so, the cuopt.so and the header files for the project all of which are installed in `/usr/local/lib` in a standard fashion. 

The headers are in `/usr/local/lib/cuopt`

Note: There are items in the deb package creation that I took the liberty to fill, some of them may need changes. 
Specifically these two lines:
`set(CPACK_DEBIAN_DEV_PACKAGE_MAINTAINER "NVIDIA")`
`set(CPACK_DEBIAN_RUNTIME_PACKAGE_MAINTAINER "NVIDIA")`
I assumed that is how you'd like the maintainer referenced.

Additionally the deb packages do not correctly reference their dependencies yet, i'm happy to follow up with that work if this looks likely to be merged.

## Issue
#189

Authors:
  - Mark Ostroot (https://github.com/mostroot)
  - Ramakrishnap (https://github.com/rgsl888prabhu)

Approvers:
  - Trevor McKay (https://github.com/tmckayus)

URL: https://github.com/NVIDIA/cuopt/pull/190