FP as a local search (#258)
This PR adds a new local search operator: FP with objective cutoff.

We were using FP as a means of solution generation algorithm. Now we are also using it as a local search by adding a cutting plane at the current fesible point perpendicular to the objective. We rerun FJ at this point to have a different restart point and continue the FP search with the cutting plane added.

This on brings down the average optimality gap by 1-2%.

Authors:
  - Akif ÇÖRDÜK (https://github.com/akifcorduk)

Approvers:
  - Alice Boucher (https://github.com/aliceb-nv)

URL: https://github.com/NVIDIA/cuopt/pull/258