Heuristic Improvements: balance between generation and improvement heuristics (#382)
This PR changes the heuristic structure by creating a natural balance between generation and improvement.
The FP/FJ loop now adds solution to the population and only if we have enough diverse solutions we exit the loop and execute the population improvement. The diversity is increased to `sqrt(n_integers)`. The recombiners are run between the current best and all other solutions in the current population, if stagnation is detected in FP/FJ loop and then the loop continues. The bounds prop rounding in the context of FP is also improved. When the dual simplex solution is set, the pdlp is warm started now with both primal and dual solutions.

The default tolerance is now 1e-6 absolute tolerance and 1e-12 relative tolerance.

This PR includes bug fixes on:
- Apperance of inf/nan on `z` vector dual simplex phase2.
- Invalid launch dimensions on FJ and hash kernels.
- Timer diff and function time limit issues when the solver is run with unlimited time limit.

Benchmark results in 10 mins run on H100:
- Main branch: 207 feasible solutions and average gap: '28.54', 3 unfinished/crashed
- This PR: 213 feasible and average gap: '23.11', 1 unfinished/crushed. (The PR didn't have any crash before merge with main branch)

closes #142  
closes #374 
closes #218

Authors:
  - Akif ÇÖRDÜK (https://github.com/akifcorduk)

Approvers:
  - Ramakrishnap (https://github.com/rgsl888prabhu)
  - Alice Boucher (https://github.com/aliceb-nv)

URL: https://github.com/NVIDIA/cuopt/pull/382