Fix an UMA on ns1952667 due to ignoring rank in factorize_basis when no singletons present (#470)
When a basis had no singletons we were incorrectly setting the rank to m, even if the right-looking LU reported that the basis was rank deficient. This caused the column pointers and row indices in the L factor to be invalid. (Since the L factor itself was invalid).

We now correctly handle rank deficiency in the basis when there are no singletons.

Authors:
  - Chris Maes (https://github.com/chris-maes)

Approvers:
  - Hugo Linsenmaier (https://github.com/hlinsen)

URL: https://github.com/NVIDIA/cuopt/pull/470