Improve Numerical Stability in Dual Simplex (#591)
This PR tighten the tolerance for $\mu = 1 + v^T u$ that indicates when to refactor the basis in dual simplex, improving the numerical stability of the method. This PR also disables the logs in the RINS heuristic.

The results from `netlib` remains unchanged in terms of number of iterations, time and objective value for `simplex` and `pdlp + crossover` and `barrier + crossover`.  The average gap for MIPLIB is comparable with the main branch, i.e., $13%$ with 226 feasible solutions on a Grace-Hopper node. This solves the numerical issues found in `b1c1s1` and `p200x1188c`.

This PR also fixes the occasional crash caused by selecting the incorrect last column when updating the basis.

Closes #593. Closes #585.

Authors:
  - Nicolas L. Guidotti (https://github.com/nguidotti)

Approvers:
  - Rajesh Gandham (https://github.com/rg20)
  - Chris Maes (https://github.com/chris-maes)

URL: https://github.com/NVIDIA/cuopt/pull/591