Propagate the bounds from the parent to the child nodes (#473)
With this PR, the solver propagates the bounds from the parent to the child nodes. If the solver reaches the end of a branch (i.e., when the node becomes infeasible or an integer feasible solution is found), the bounds are recomputed from scratch. This allows the solver to reuse the tightened bounds from the parent and re-apply the bounds strengthening on the branched variable.


Average (Primal) Gap over the MIPLIB2017 dataset: 
`main` (10f116b): `14.472358` with `224` feasible solutions
This PR: `14.058718` with `226` feasible solutions

On average, the `main` branch explored `1012412` nodes. This PR increases to `1607821`  nodes (i.e., `58%` more nodes) for the same time frame.

Authors:
  - Nicolas L. Guidotti (https://github.com/nguidotti)

Approvers:
  - Rajesh Gandham (https://github.com/rg20)
  - Chris Maes (https://github.com/chris-maes)
  - Akif ÇÖRDÜK (https://github.com/akifcorduk)
  - Ramakrishnap (https://github.com/rgsl888prabhu)

URL: https://github.com/NVIDIA/cuopt/pull/473