Reuse of the basis factorization from the parent node (#492)
With this PR, the child node can reuse the basis factorisation from the parent, reducing the time required for solving the LP relaxation of the child. The solver recomputes the basis factorisation when the branch reaches the end. 
This is a follow-up of #383 and #473. 

Average (Primal) Gap over the MIPLIB2017 dataset: 
`main` (10f116b): `14.472358` with `224` feasible solutions
This PR: `~14.26` with `225` feasible solutions

On average, the `main` branch explored `1012412` nodes. This PR increases to `1850797`  nodes (i.e., `82%` more nodes) for the same time frame. Compared to #473, it explores `15%` more nodes.

This also fixes an incorrect report of an infeasible solution after a timeout and there are no nodes in the heap.

Authors:
  - Nicolas L. Guidotti (https://github.com/nguidotti)
  - Chris Maes (https://github.com/chris-maes)
  - Alice Boucher (https://github.com/aliceb-nv)

Approvers:
  - Chris Maes (https://github.com/chris-maes)

URL: https://github.com/NVIDIA/cuopt/pull/492