Fix infinite average pseudocost + Handle numerical issues in strong branching (#678)
If the LP relaxation when branching on a fractional variable in the strong branching do not lead to a optimal solution and the iteration limit is not reach, then the corresponding pseudocost is set to infinite. This ignore any issue in the dual simplex and cause the pseudocost average to also be infinite.

This PR properly handle numerical issues in the dual simplex in strong branching and exclude infinite pseudocost to be included in the average.

Not critical for 25.12.

Authors:
  - Nicolas L. Guidotti (https://github.com/nguidotti)

Approvers:
  - Chris Maes (https://github.com/chris-maes)

URL: https://github.com/NVIDIA/cuopt/pull/678